Add option to make debugging queries private
Add an option to make debugging queries private, meaning only
accepted from the loopback address. The debugging query
(i.e. "_status" right now) contains
information which may be either commercially sensitive or
security sensitive. Therefore provide an option (-privatedebug)
which if set does not return these unless the query comes
from the loopback address.

"_country" is not so restricted as its only use for
debugging is from a non-loopback IP.

The logic looks a little clunky - this is to permit easy
merging of the "_health" RR later.

Signed-off-by: Alex Bligh <alex@alex.org.uk>